Use github.com/containers/image to get image digests
It appear the way we were looking for image digests might not work with
some registries because of different docker registry references used.
Also getting the digest with signed registries is quite complex.
So use github.com/containers/image which provide the proper way for
working with registries.

We lost auto-detection of imagePullSecrets, but I think it's ok to let
the user mount the secret to have a ~/.docker/config.json (so RBAC
doesn't need to have access to Secrets).

Closes #13